Factor out search box into a module
